spline and influences added, not done
currently working on event system for displaying the influences

+++ Assets/Scripts/final/Influence/Influence.cs
@@ -3,12 +3,12 @@ using UnityEngine;
 public class Influence
 {
     public Vector3 Position { get; }
-    public int Radius { get; }
+    public float Radius { get; }
     public string PromptModifier { get; }
 
     public GameObject Prefab { get; }
 
-    public Influence(Vector3 position, int radius, string promptModifier, GameObject prefab)
+    public Influence(Vector3 position, float radius, string promptModifier, GameObject prefab)
     {
         Position = position;
         Radius = radius;
